surrogate/backend/pyproject.toml
[tool.poetry]
name = "image-app-backend"
version = "0.1.0"
description = "Backend für Image Application mit OpenClip"
authors = ["Your Name"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
numpy = "^1.26.4"
torch = "^2.2.2"
scipy = "^1.13.0"
open_clip_torch = "^2.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"